Re-added integration test harness.
This harness used to be in XDMoD and was not copied across when the
repo was originally made public.

Also include a couple of example test files that check the metric
explorer and usage explorer controllers. These files are intended to
be used as templates to help developers write their own integration
tests.