Added ability to link directly to jobs.
Right clicking on a job in the search history tree now gives a context
menu with the option to get a sharable link. This URL can be used to
directly link to a given job for all users of the system (although the
job permissions are not changed, so only those users that have
permission will be able to actually view the job data.

Job search based on the local job id is also possible, but requires the
knowledge of the primary key for the resource that ran the job. This
feature is intended for developer use (e.g. to link to the job viewer
from an external system).