Ensure assist panel shows when there are no job tabs.
The assist panel did not show when the job tabs were removed due to a
right-click delete search in the search history tree. This was because
the tab.removeAll() function did not fire the close event. Changed the
code to listen to the destroy event which is always fired.

Steps to test:

- Open one or more jobs in the job viewer
- in the searhc history tree right click on a search node and select
  'Delete Search'
- Click 'yes' when prompted

With the original code the main tab panel is blank. This change results
in the assist panel being displayed again.