Added number of filters selected on Data Series Definition dialog.
Created a new function called numberOfFiltersSelected to loops through the filters store and look to see how many of the records are marked as checked. 

There is a getSelectedFilters function that returns all the filters in the store and how many there are but it does not check to make sure they are selected or not. These two functions can be merged into one if we want.