Fix incorrect data filtering
Fixes issue with data filtering when an unpriv user tries to view
timeseries data for a realm / stat that is protected (i.e. has
additional role-specific filters applied).

Added integration test that confirms the correct behaviour.

Note that this resolves the regression issue created in the (reverted)
pul request #387. The full regression tests have been run on this code
change.