Simplifying the backup / restore of unmanaged tables
Having the sql inline w/ the code obfuscated the purpose of the code and made it
harder to read. To address this problem each of the sql statements were moved
out to their own commented sql file.

The names of the newly added functions were updated to make their purpose more
intuitive.

A few of the log messages had their contents simplified.