Renaming deriveUserEnumerationQuery function
The results of this function have changed significantly and as such a renaming
is in order. We are no longer deriving a User Enumeration Query but instead
listing user emails by group and acl.

The `role_filter` parameter has also been renamed `acl_filter` to more
accurately represent what it contains.