Test Additions: Pre-Remove Role Usage outside of XDUser (#478)
* Spacing Updates

Adding test for getUserVisits

- This test exercises the:
    `internal_dashboard/controllers/controller.php?operation=enum_user_visits`
  path.
- This is path was chosen because of the use of Role* tables in the function
  that retrieves the data for this operation: `XDStatistics::getUserVisitStats`.

Updated TestFiles to allow the caller to specify file type

- Added a new parameter w/ a default value `$fileType = '.json'`. This new
  parameter will, if specified, allow the caller to retrieve something other
  than json files ( while keeping the previous behavior the same ).

Adding a test for enum_user_visits_export

- This operation utilizes the same sql function that enum_user_visits with the
  addition of the output being returned as a `.csv`

Adding a test for operation enum_target_addresses

- This operation is utilized by the Batch Email modal provided by the 'Prepare
  Email' button in the 'User Management' => 'Existing Users' tab.
- Currently it returns both the 'pub' and 'admin' users. It is likely that these
  will be removed from the results when we migrate the SQL from Role* to Acl*.
  Ensure that new expected results are generated / added after the modification
  is made.

Initial Tests for XDChartPool & XDReportManager

Adding tests that exercise endpoints in /controllers/report_builder

The tests added in this commit exercise the endpoints that utilize
`getFormalRoleNameFromIdentifier`, which queries Role* tables, via
`XDChartPool::enumChartsUnderOtherRoles` and
`XDReportManager::enumReportsUnderOtherRoles`.

Updating removeUser success logic

Previously we expected that the the 'success' property returned by the
remove_user operation was always going to be true and that the user name of the
user being removed would be in the message property. This did not account for
the possibility of requesting a user be removed that does not exist. To resolve
that oversight the logic is now:
  - if the message contains 'user_does_not_exist' then the 'success' property
    must be false.
  - else, the 'success' property must be true and the user name must be found in
    the 'message'.

I also added a try-catch around the calling of removeUser w/ information that
helps to debug any problems there may be with the execution of the removeUser
function.

Migrated the ReportGeneratorTest tests to ReportBuilderTest

I moved the one test in ReportGeneratorTest to ReportBuilderTest as there's no
need to have it separated out.

Adding tests for UserControllerProvider

This test exercises the users/current endpoint. Specifically as it utilizes the
aRole::getFormalName function which queries the Roles table.

Broadened testEnumTargetAddresses test cases

- Added additional test cases that covered empty, null, large text and large
  numeric test cases. This required a slight re-organization of the test data
  and subsequently, a re-organization of the code retrieving said test data.
- Added a new TestHelper that can, based on the format of a test parameter
  value,  generate random length strings / numbers for use as test data.

Adding the an explicit test for the create_chart endpoint

Previously the create_chart endpoint was only tested as a part of the
create_report endpoint. This new test exercises the create_chart outside of
creating reports. Specifically, it includes scenarios that we expect to not be
successful.

Reorganizing the createChart / createReport tests

Having the `createChart` tests declared ahead of the `createReport` tests
introduced situations in which `createReport` would fail. By having it declared
after `createReport` these failures do not occur.

* Re-applying changes per code review comments by @jpwhite4

* Applying suggestions per code review by @smgallo

* Applying suggestions from code review by @jpwhite4

* Fixing misspelling in comments

* Updating the provideEnumTargetAddresses

Previously we were not processing the test data parameters as expected. The code
has been updated to properly handle the structure of the incoming data.

* Add a test to ensure getUserVisits increments correctly

per a conversation with @jpwhite4 I'm adding a test that ensures the user_visits
endpoint correctly increments a user's visits.

* Update the getUserVisitsExport test

- Added documentation to various portions of this function to help clarify
  intent.
- Updated the way the actual output is processed to ensure that we're comparing
  apples to apples.
- Updated the algorithm used to determine if the actual output matches the
  expected.

* Adding a helper function `getUserVisits`

Since we end up calling this endpoint quite a lot I've added a helper function
that takes care of the validation of the response and the returned data.

* Adding function doc-blocks

* Removing an extraneous empty line.

* Separating out UserAdminTests

The UserAdminTests that were creating / modifying users were causing problems
with those that don't ( particularly the enumUserVisitsExport tests ). By
splitting these out we resolved the problem.