Updating ProfileRoleDelegation to handle demotion correctly
Prior to this commit if a center director demoted a user who was only associated
with their center via a Center Staff Acl the UI would incorrectly keep the user
in the list of eligible users & if selected would report a center_mismatch.

This arose due to the UI not accurately interpreting the back end state.
Specifically, it is expected that if a user, as described previously, is demoted
that they would no longer be included in the list of users eligible for
promotion to Center Staff as they are not currently associated with the center.
So it should not be possible to see them listed let alone select them.

These changes now ensure that each time a user is demoted that the list of
eligible users is refreshed and if the demoted user is found in the newly
refreshed list, that user is selected. Else the user selection is cleared and
status message reset to it's value upon initial viewing of the promotion ui.