Defaulting module value for local configs
There are cases when we want to break out configuration data but  not associated
it with a module. Previously this data would not have been included / processed,
but now if the module of a local config file is not known then `xdmod` is used.

In particular this resolves a problem that arose due to the way the Cloud
configuration files are organized ( at the time of this commit ). This was
preventing them from being properly interpreted by XDMoD. These changes allows
these files to be interpreted correctly.