Adding Organization Population to SSO
Attempting to retrieve / set the organization for this user upon acct creation.

Updating the new user notification to include organization info

Adding some documentation to the notifyAdmin function

Just wanted to clarify what's going on / why.

Adding email logging for admins & new sso users

Added an email logger that will notify admin users that additional setup may be
required if the users organization was not found but an organization_id property
was provided by SimpleSAML.

Also added the sending of an email to the user whose organization was not able
to be found informing them that additional setup may be required.

Adding a fallback if organization_id is not found

Just adding a fallback so that if an organization cannot be found utilizing
`organization_id` then attempt to use `organization`.

Updated unableToFindOrganization variable

Added checking for the organization saml attribute to whether or not we were
unable to find an organization.

Increasing the robustness of notifyAdminOfNewUser

MailWrapper::sendMail can throw an exception if it is unable to successfully
execute sendmail. By wrapping the statement in a try-catch we allow the
application to continue functioning while still logging that we were unable to
notify the SSO user further steps need to be taken.

Updates per code review by @jpwhite4

Updates per @jpwhite4 code review comments

While we always receive these attributes in our environment, this is wholly
dependent on what the IDP sends back. By moving organization's if-else block out
of organization_ids we allow for organization to be found regardless of whether
or not organization_id is provided.

Updates per @jpwhite4 code review comments

Explicitly defined the UNKNOWN_ORGANIZATION via it's id

re-removing the UNKNOWN_ORGANIZATION constant

Per the replace_role_parameters PR we're just using -1 instead of constants.

Updated to include renamed function name

Updating default organization result

Updated default organization_id == -1

Updating Default Values

These values were previously `null`, just changing them to `-1` to be
consistent.

Removing check for unneeded SAML attribute

Per code review comment by @chakrabortyr we don't need to check for 'email' as
it's not reliably populated.

Add configuration option for Admin Notification

Added a new configuration setting that allows Admins to opt-in/out ( it's on by
default ) of receiving emails when an SSO User logs in but we are unable to
identify which organization they are associated with.

Code cleanup

The `notifyAdminOfNewUser` function really notified the admin AND the user. I've
extracted the user notification portion into it's own function and slightly
expanded the original function name so that it's more clear on what it's doing /
when it should be called.

Updating portal_settings template

Forgot to update the template.

Include json content as opposed to PHP for the admin email

Updating XDSamlAuthentication to use new config settings

We now have 3 configuration settings ( portal_settings.ini ) that help to
control what happens when a new SSO user logs in. These settings are:
- `default_organization_name`: The name of the organization to use when
  `force_default_organization` is `on`
- `force_default_organization`: When `on` all new SSO users will be associated
  with the organization identified by `default_organization_name`.
- `email_admin_sso_unknown_org`: When `on` the XDMoD Admins will be notified via
  email when a new SSO user logs in and, they either do not have organization
  information or their organization information is not found in XDMoD

Organizations::getOrganizationIdByLongName: removed as it is no longer used.

Add additional setup information

Updated the Organization Setup so that we can use the specified organization
name in portal_settings.ini as `default_organization_name`.

Added the new configuration settings to portal_settings.ini and
portal_settings.template

Updated setup scripts

samlSetup.sh:
- We're going to be standardizing on the SAML attribute `organization` which
necessitated changes to the SAML test data.

xdmod-setup.tcl:
- Needed to accommodate the additional prompt during organization setup

Updating portal_settings files

Forgot to add the section name to the `default_organization_name` variable. This
was causing it to not be populated correctly and thus causing the unit tests to
fail.

Also needed to remove `internal.email_admin_sso_unknown_org` once it was moved
to the `[sso]` section.

Cleaning up unused functions