Updated enum_user_types_and_roles to utilize acls (#446)
* Updated enum_user_types_and_roles to utilize acls

- Also updated mailer.php:enum_target_addresses as it utilized the results of
  enum_user_types_and_roles.

* Updating the test artifact for testEnumUserTypesAndRoles

- This controller now returns acl_ids as opposed to role_id so the values needed
  to be updated for this branch.

* Adding 'AS' keyword per comment by @jpwhite4

* Simplifying SQL / Removing Unneeded code

We can remove the 'ORDER BY' clauses as we're sorting the results in PHP after
retrieval.

We can remove the `array_unique` as we have the `DISTINCT` keyword in the sql
query.

* Renaming deriveUserEnumerationQuery function

The results of this function have changed significantly and as such a renaming
is in order. We are no longer deriving a User Enumeration Query but instead
listing user emails by group and acl.

The `role_filter` parameter has also been renamed `acl_filter` to more
accurately represent what it contains.

* Updating input for test affected by this PRs changes

enum_target_addresses was affected by the changes introduced in this PR so the
input file has been updated to reflect which new output files to utilize during
testing.

* Pre-Change Cleanup

Just cleaning up the file from a style standpoint. Changes include:
- normalizing the spacing @ 4
- renaming snake case variables to camel case
- removed unnecessary comment lines ( i.e. they contained no actual comments )
- removed some leading and trailing empty new lines within code blocks
- removed the trailing ?> as it is unnecessary