Update Center Promotion (#509)
* Spacing Updates

* Spacing Updates

* Adding integration tests for role delegation behavior

* Updating role_manager operations and supporting classes

Previously User Role Delegation worked as follows:
- A Center Director could delegate the Center Director Role to Center Staff
- A Center Director could demote Center Directors that they had promoted back to
  Center Staff.

Currently User Role Delegation works as follows:
- A Center Director can delegate Center Staff to any user "associated" with
  their center.
  - "associated" means that the user either has a direct relation with the
    center ( via Users.organization_id ) or is mapped to a person that has a
    direct relation with the center ( via Users.person_id -> Persons.id ->
    Persons.organization_id )
- A Center Director can demote a Center Staff "associated" with their center by
  revoking the previously granted access to their center. If the user has no
  other centers they are Center Staff of, then the Center Staff acl is removed.

**NOTE: There are accompanying changes to how Federated Users are created
        that will need to be added to ensure that they are associated with a
        center / organization.**

* Updating the sql for users eligible for promotion

- Updated to take into account the correct organization fallback steps i.e.
  - If the user has a user_acl_group_bys record that has a value that is also
    a valid center.
  - If the user is directly related to a valid center ( Users.organization_id )
  - If the user is related to a valid center via their assigned Person.
- Updated the function parameter accordingly ( we now require the user id of the
  person running the query.

* Updated additional sql queries to take into account center fallback

- Just needed to add the center fallback queries to these two sql queries.

* Pre-Change Cleanup

Cleaning up the file stylistically before making any functional changes.

* Updating ProfileRoleDelegation to handle demotion correctly

Prior to this commit if a center director demoted a user who was only associated
with their center via a Center Staff Acl the UI would incorrectly keep the user
in the list of eligible users & if selected would report a center_mismatch.

This arose due to the UI not accurately interpreting the back end state.
Specifically, it is expected that if a user, as described previously, is demoted
that they would no longer be included in the list of users eligible for
promotion to Center Staff as they are not currently associated with the center.
So it should not be possible to see them listed let alone select them.

These changes now ensure that each time a user is demoted that the list of
eligible users is refreshed and if the demoted user is found in the newly
refreshed list, that user is selected. Else the user selection is cleared and
status message reset to it's value upon initial viewing of the promotion ui.

* Forgot to actually kick off the reload

* Changes per code review comments by @smgallo

* Missed a db declaration

* Additional Changes per convo w/ @smgallo

* re-adding schema to from clauses

Just replacing the schemas for a few from clauses so that they're all explicitly
declared.