Removing invalid package-lock.json
The hash for the xmldom package specified in package-lock.json doesn't match the
hash generated by the package that is downloaded. This meant that the saml idp
is never setup. Which means that the saml tests fail. By removing the lock file
we allow npm to generate a new one / finish installing / the tests to pass.