Migrating aRole functions to acls
Added two functions to the `Acls` service class that reproduces the behavior /
results of their aRole counterparts but utilizes the new acl system. The
functions replaced are:
  - `getQueryDescripters`
  - `hasDataAccess`

Both functions have had comprehensive component tests that checks that their
output exactly matches their aRole counterparts.

There is one big difference between the `aRole` functions and the `Acls`
functions. The `Acls` functions take all of a users acls into consideration when
generating their output whereas the `aRole` functions filtered their output on
the role it represented. This caused a problem when defining a new non-feature (
i.e. it provides data access / filtering ), non-hierarchical acl. Specifically,
the new acl would not be taken into consideration when populating the realms /
group bys in the user interface. This was due to the source of that data being
the users most privileged role.