Updating the demangle function
I ran into a problem with this function outputting invalid json which
subsequently caused tests to die when they attempted to utilize the invalid
decoded json. After investigation I found that the reason was due to
constructing invalid string values. The invalid json would look something like:

```json
{
  "key1": "value1",
  "key2": "value2",
  "key3": ""FUNCTION DELETED""
}
```

Notice the 2 x double quotes surrounding 'FUNCTION DELETED', this was the root
of the problem. I just added some code to determine whether or not to
encapsulate the replacement message in quotes or not.