Updated sql for getDisabledMenus && getQueryDescripters
The sql for these two functions have been updated so that the only acl records
utilized are:
  - The most privileged acl per hierarchy
  - any non-hierarchical data acls
The reason for this change came about during XSEDE testing where there is a much
more complicated authorization setup.

The `getQueryDescripters` function has additional changes so that it correctly
handles the presence of a provided statisticName. The presence of statistic
values in the sql when no statistic was requested was causing non-intuitive
results to be returned.