Support Globalize Query Descripters (#560)
* Globalize Query Descripters (#552)

* Basic Documentation Cleanup

Just cleaning up the documentation ahead of committing actual code changes.

* Documenting Confusing Code

After tracing back through the code I found that these first two conditions can
never actually be met. Just Adding some documentation here so that future devs
don't have to go through the same thing.

* Migrating aRole functions to acls

Added two functions to the `Acls` service class that reproduces the behavior /
results of their aRole counterparts but utilizes the new acl system. The
functions replaced are:
  - `getQueryDescripters`
  - `hasDataAccess`

Both functions have had comprehensive component tests that checks that their
output exactly matches their aRole counterparts.

There is one big difference between the `aRole` functions and the `Acls`
functions. The `Acls` functions take all of a users acls into consideration when
generating their output whereas the `aRole` functions filtered their output on
the role it represented. This caused a problem when defining a new non-feature (
i.e. it provides data access / filtering ), non-hierarchical acl. Specifically,
the new acl would not be taken into consideration when populating the realms /
group bys in the user interface. This was due to the source of that data being
the users most privileged role.

* Updating the callers of the updated functions

Just changing the callers of the updated functions to utilize the new functions.

* Small travis style fix

* Remove unnecessary code

Sooooo, it turns out that we do have the required information in the database so
there's no need to do the crazy merge from roles.json thing. This commit removes
the complicated merge code and adds the two sql columns that replace it.

* Updating Acls::getQueryDescripters

- removed the `visible` column as it's only used when filtering output and the
  setting of the `show_menu` query descripter property as it's not used here.
- changed `\`disable\`` to `not_enabled` so that we don't have to use back
  ticks.
- Added some group bys so that we can get rid of the `DISTINCT`

* Updated tests to use expected test artifacts

- Updated the newly added tests so that they use test artifacts as opposed to
  checking the old method output vs new method output.
- This is really so that we can remove the old ( and now unused ) methods.

* Removing getDisabledMenus

- This function is no longer used so it's being removed.

* Removing aRole::hasDataAccess

- This function is no longer being used and as such is being removed.

* Updating usage of getQueryDescripters

Just updating QueryBuilder to use the new `Acls::getQueryDescripters` function.

* Removing unused function

[a|i]Role::getQueryDescripters is no longer used so it's being removed.

* Updates per code review by @smgallo

* Updates per code review by @jpwhite4

* Updates per code review comments by @jpwhite4

* Updates per code review comment by @jpwhite4

* Adding tests artifacts

Just adding some test artifacts that were added in between the moving of the
artifacts into the main repo.

* renaming output files appropriately

* removing unrelated test artifacts

* Making the testGetQueryDescripters test more robust

Swapped out the serialized php expected files for json that only contains the
information that we're interested in comparing.

* Updating testGetQueryDescripters

Updated the  test to utilize the new json test artifacts

* re-adding required test artifacts

* Adding jobwaittime to test artifacts

Needed to add jobwaittime to get_query_descripters test artifact.

* Adding required function

Just adding a function that's utilized by the tests to the BaseTest

* Updating the demangle function

I ran into a problem with this function outputting invalid json which
subsequently caused tests to die when they attempted to utilize the invalid
decoded json. After investigation I found that the reason was due to
constructing invalid string values. The invalid json would look something like:

```json
{
  "key1": "value1",
  "key2": "value2",
  "key3": ""FUNCTION DELETED""
}
```

Notice the 2 x double quotes surrounding 'FUNCTION DELETED', this was the root
of the problem. I just added some code to determine whether or not to
encapsulate the replacement message in quotes or not.

* Updated sql for getDisabledMenus && getQueryDescripters

The sql for these two functions have been updated so that the only acl records
utilized are:
  - The most privileged acl per hierarchy
  - any non-hierarchical data acls
The reason for this change came about during XSEDE testing where there is a much
more complicated authorization setup.

The `getQueryDescripters` function has additional changes so that it correctly
handles the presence of a provided statisticName. The presence of statistic
values in the sql when no statistic was requested was causing non-intuitive
results to be returned.

* Updated expected output for sem_* statistics

None of these statistics should have been visible per settings in
datawarehouse.json. The expected output has been updated accordingly.

* Adding password fallback to the getUserbyUsername tests

Just adding in a fallback for users that still have passwords hashed via the old
method.

* Making sure that acl-config runs on xdmod-update