Updating to account for removal of UNKONWN_ORGANIZATION_NAME
Just updated the sql for this function so that there will always be at least one
record returned ( if no organization is found for the provided `organization_id`
then the unknown organization name will be returned ).

Also, since we are now guaranteed to have at least one record we can simplify
the return clause.