Grand Cleanup and Refactoring of Cloud Realm (#612)
<!--- Provide a general summary of your changes in the Title above -->
This PR should fail until the anonymized test data is merged. It should also not be merged until  Shippable uses the regression tests we've added.

## Description
As the title suggests. Changed names + descriptions for clarity and consistency with the rest of XDMoD. Changed default/unknown values to align with what we do in other realms and updated data types to reflect this. 

## Motivation and Context
Consistency, actually having tests

## Tests performed
Tests were added in this branch. Also manual audits of the data on my VM and in metrics-dev.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [X] My code follows the code style of this project as found in the **CONTRIBUTING** document.
- [X] I have added tests to cover my changes.
- [ ] All new and existing tests passed.