Update the samlSetup script to support running on XSEDE (#630)
* Updated the samlSetup script

This script can now be used in the xsede docker as well as the
OpenXDMoD docker

* Updates per code review comment by @jpwhite4

* re-added the `DEFAULT_` variables

We need to be able to look for the values contained in `DEFAULT_INSTALL_DIR` /
`DEFAULT_VENDOR_DIR` in `xdmod.conf` for replacement. This is independent of
the values specified for the `INSTALL_DIR` / `VENDOR_DIR`.