Attempting to diagnose / resolve Shippable test problem
Running into a problem w/ Shippable ( fresh_install ) that I'm not able to
re-produce locally.

The error encountered is:
  - `An element could not be located on the page using the
given search parameters`.

This is encountered at the: `Metric Explorer` -> `Available For Report` ->
`Select the first saved chart` step and seems to be related to the
`me.selectors.load.firstSaved` property which corresponds to the first chart in
the available charts list in the Report Generator tab.

These changes attempt to ensure that the `selector` provided is visible before
clicking on it.