Updating password member variable during saveUser
There is a problem with the XDUser object state getting out of sync with it's
record in the DB. Specifically, when an XDUser is created / saved the password
in the XDUser object is the unhashed value where as the DB has the hashed value.
This change updates the member variable to be in line with the DB value.