Fix for user organization not populating (#680)
* Fix for user organization not populating

Per testing by @plessbd. When initializing more than one new user account from
requests, unless the user's 'Organization' was found by name in the db, then
when a person was selected to map the new user to the organization id would be
displayed not the organization name. The steps to re-produce are:

  - Start with at least 2 user account requests
  - Select the first user account request
  - Click 'Initialize New User Dialog'
  - Assign a Person to this User.
  - click 'cancel'
  - Select the next user account request
  - Click 'Initialize New User Dialog'
  - Assign a Person to this User.
  - Notice that the 'Institution' box displays a number instead of a name.

The reason this was occurring is that the data store used to populate the
Institution drop down was only being refreshed ( without a filter ) the first
time the dialog opened. Each time after that it was being filtered with the
organization value from the account request. So, when a Person would be chosen
to map the user to the code would attempt to automatically set the Institution
value ( via id ). but, because the data had been pre-filtered there would be no associated
value and so it would default to displaying the numeric id instead.

This change makes `enum_institutions` a little smarter in that if no filtered
records are found it instead retrieves / returns all the organizations.

* Adding documentation to the code change

Just a small description of why the new code was added.