Update logic for setting a users organization
Since we've changed the user to be directly related to an organization ( via
Users.organization_id ). We no longer provide the center a user is related to
via a specific acl. This necessitated changing the logic from checking if there
were any centers found w/ the acls to, do they have a center related acl? Do
they have an `instititution` ( organization )? If so then make sure that we call
`setOrganizations` for it. This will ensure that both
`user_acl_group_by_parameters` and `UserRoleParameters` are populated correctly.