Adding 'My Profile' UI Tests
These tests check that the information presented on the 'General' tab of the 'My
Profile' dialog are displayed as expected. In particular the 'Top Role' test
ensures that the value displayed for users w/ a center related acl ( i.e. center
director or center staff ) is in the format: `<most privileged acl> - <center
name>` where users without these acls are in the format: `<most privileged acl>`