Unlock the Institution drop down when Unknown Person is selected
We did not account for needing to manually override a User's associated
organization when they are mapped to the 'Unknown' person. These changes enable
that on the front end and make sure that they are correctly taken into account
on the back end.

Updating user facing emails per @jpwhite4

Updating to the greeting to include the user's name.