Fix Role Delegation (#685)
* Update logic for setting a users organization

Since we've changed the user to be directly related to an organization ( via
Users.organization_id ). We no longer provide the center a user is related to
via a specific acl. This necessitated changing the logic from checking if there
were any centers found w/ the acls to, do they have a center related acl? Do
they have an `instititution` ( organization )? If so then make sure that we call
`setOrganizations` for it. This will ensure that both
`user_acl_group_by_parameters` and `UserRoleParameters` are populated correctly.

* Adding 'My Profile' UI Tests

These tests check that the information presented on the 'General' tab of the 'My
Profile' dialog are displayed as expected. In particular the 'Top Role' test
ensures that the value displayed for users w/ a center related acl ( i.e. center
director or center staff ) is in the format: `<most privileged acl> - <center
name>` where users without these acls are in the format: `<most privileged acl>`

* Adding validation for the 'institution' parameter

Ensuring that the 'institution' parameter is provided for
[create|update]_user.php and that it is only composed of numbers ( optionally
negative ).

* Removing unused code

* Add newline per @plessbd comment