Unlock org for unknown person (#689)
* Unlock the Institution drop down when Unknown Person is selected

We did not account for needing to manually override a User's associated
organization when they are mapped to the 'Unknown' person. These changes enable
that on the front end and make sure that they are correctly taken into account
on the back end.

Updating user facing emails per @jpwhite4

Updating to the greeting to include the user's name.

* Changes to support new UI Tests

These changes mostly include the addition of class or id attributes. The
changes to `[create|update]_user.php` were so that we can reliably select based
on the contents returned ( i.e. the added <b> was causing problems when
attempting to select the element that contained the full message. )

Adding some basic Internal Dashboard UI Tests

These tests include basic ones for creating and removing a user.