Updating Sticky Checkbox functionality
Per convo w/ @jpwhite4: Previously the sticky checkbox would only update on
`blur` ( `change` ) but this led to a confusing user experience. This has been
updated so that on `select`, `blur` or `change` if the user mapping or
organization association has changed then the sticky checkbox is checked, else (
it's the same as when the user was loaded ) go ahead and un-check it.