Refactor / Simplification of Usage Filter generation
Refactored a portion of the `convertChartRequest` function and combined the two
`for-each` loops that were populating the `$meFilters` variable. Now we have one
loop that has the same functionality.

Also a tiny comment update in the `translateFilterValue` function for the
'pi' filter type. Now that we always return an array the previous comment didn't
really make sense anymore.