Replacing shell_exec w/ proc_open
This PR is the result of addressing the concerns brought up in support ticket:
https://help.xdmod.org/a/tickets/11402

Unable to differentiate between a command that has no output and an command that
errors out when using `shell_exec`. By replacing this with `proc_open` we're
able to retain STDOUT and STDERR but we also get access to the commands
return value. We make sure to not swallow when a command errors out by throwing
an exception when this is detected.