Acl script consolidation (#667)
* Acl Script Cleanup / Consolidation

These changes ensure that the only script that needs to be executed to bring
things up to date w/ Acls is `acl-config`.

Since `acl-config` now handles the execution of `acl-xdmod-management` and
`acl-import`, we can remove the scripts themselves and any references to them in
other code.

* Removing unused code

* Adding Documentation

* Adding the final drop tables back in

This was removed during the previous rebase operation. Just adding it back in.

* Adding Additional Documentation / logging

Just adding some additional documentation / logging for the new functions.

* Updating Admin User Setup

Since users are now directly associated with organizations we need to retrieve
the organization associated with this XDMoD Install and associate the admin user
with it. If unable to retrieve / lookup the organization then default to the
'Unknown' organization.

* Updating the docker image file for shippable

The latest image includes the updated user creation scripts ( for relating users
to an organization ).

* Changing dockerfile back to v5

* Updating to utilize Etl* classes per convo w/ @plessbd

* Removing unused code

This class is no longer used and as such it's being removed.

* Reverting admin user creation

There's no need for admin to be associated with the organization associated with
the XDMoD installation. The "Unknown" organization is perfectly valid.

* removing unused code

This function is no longer called.

* Updating acl documentation

`acl-config` is now the one stop shop for all things acl. As such we no longer
need to have separate scripts for `acl-import` / `acl-xdmod-management`. These
change the documentation to reflect this.

* Doc updates per @jpwhite4

* Update Java version in RPM list

* Passing `dryRun` to EtlOverseer

Per @jpwhite4, letting ETLOverseer handle `dryRun`.