Revert workaround for thumbnail slowness.
This patch reverts commit 43fa0377154e35ec639968ffefe95c88f38186b4
from Wed Nov 25 10:12:49 2015 -0500.

Prior to this change, the Usage tab in XDMoD displayed a thumbnail view
by default. The thumbnail view showed multiple charts, one for each of
the statistics of the first node in the catalog.

The XDMoD code was changed to use the metric explorer backend to generate
the data for the Usage tab. Unfortunately, the changes resulted in terrible load
performance for the thumbnail view. A short term workaround was
implemented in commit 43fa0377154e35ec639968ffefe95c88f38186b4. This
workaround loaded one of the charts by default rather than the thunmbnail
chart view.

The changes in https://github.com/ubccr/xdmod/pull/750 mitigate
the slowness so this pull request reverts the workaround.