Move store migration code from rest endpoint to migration class.
The store migration code currently runs every time the store
data is requested. (There is a flag set to only perform the migration
once per user). This means that the migration only happens when
a user logs in and ther has to be extra code in the summary tab to handle
the case where a user has not yet been migrated.